+++ apps/web/components/previews/counter.tsx
@@ -1,7 +1,16 @@
 "use client";
 
+import { AbsoluteFill } from "remotion";
 import { Counter } from "../registry-exports";
 
 export const CounterPreview: React.FC = () => (
-  <Counter from={0} to={100} suffix="%" durationInFrames={90} />
+  <AbsoluteFill
+    style={{
+      backgroundColor: "#0f172a",
+      justifyContent: "center",
+      alignItems: "center",
+    }}
+  >
+    <Counter from={0} to={100} suffix="%" durationInFrames={90} />
+  </AbsoluteFill>
 );